cleanup topk and bottomk impl.
